+++ CHANGELOG.rst
@@ -17,6 +17,7 @@ Version 2.1.0 - Unreleased
 * Updated Makefile to use colors;
 * Upgdaded package setup to split dev dependencies on extra requirements;
 * Upgraded package base dependancies to the latest versions;
+* Changed command ``init`` so the argument ``name`` is now a prompt option;
 
 
 Version 2.0.1 - 2023/08/18